requirements.txt
requests
python-dotenv
openai
pytest
phpserialize
Pillow